Improve treatment of _Optional in NullabilityChecker
No longer return from NullabilityChecker::checkPostStmt
early in the case of an explicit cast of an expression
with _Optional-qualified referenced type to a type with
Nullability::Unspecified (which should never happen for
a cast to another optional referenced type). This allows
a transition to Nullability::Contradicted to be added if
appropriate.

This is only a change to the semantics of _Nullable when used
in conjunction with _Optional, in which case the qualifier
overrules the attribute.

For example:

void foo(_Optional int *_Nullable i)
{
  int *j;
  j = (int *)i; // cast to avoid constraint violation
  *j = 5; // no diagnostic because *i is _Optional
}

For now, there is no special treatment of similar
assignments of _Optional-qualified referenced types
without cast in NullabilityChecker::checkBind, because
those are already constraint violations.